move schema to reusable location (importable by lotus/conformance).
