+++ gen/builders/generator.go
@@ -13,6 +13,8 @@ import (
 	"runtime/debug"
 	"strings"
 	"sync"
+
+	"github.com/filecoin-project/test-vectors/schema"
 )
 
 // Generator is a batch generator and organizer of test vectors.
@@ -44,7 +46,7 @@ type Generator struct {
 var GenscriptCommit = "dirty"
 
 // genData is the generation data to stamp into vectors.
-var genData = []GenerationData{
+var genData = []schema.GenerationData{
 	{
 		Source:  "genscript",
 		Version: GenscriptCommit,
@@ -55,7 +57,7 @@ func init() {
 	genData = append(genData, getBuildInfo()...)
 }
 
-func getBuildInfo() []GenerationData {
+func getBuildInfo() []schema.GenerationData {
 	deps := []string{"github.com/filecoin-project/lotus", "github.com/filecoin-project/specs-actors"}
 
 	bi, ok := debug.ReadBuildInfo()
@@ -63,12 +65,12 @@ func getBuildInfo() []GenerationData {
 		panic("cant read build info")
 	}
 
-	var result []GenerationData
+	var result []schema.GenerationData
 
 	for _, v := range bi.Deps {
 		for _, dep := range deps {
 			if strings.HasPrefix(v.Path, dep) {
-				result = append(result, GenerationData{Source: v.Path, Version: v.Version})
+				result = append(result, schema.GenerationData{Source: v.Path, Version: v.Version})
 			}
 		}
 	}
@@ -77,7 +79,7 @@ func getBuildInfo() []GenerationData {
 }
 
 type MessageVectorGenItem struct {
-	Metadata *Metadata
+	Metadata *schema.Metadata
 	Selector string
 	Func     func(*Builder)
 }
