+++ gen/suites/actor_creation/main.go
@@ -6,6 +6,7 @@ import (
 	"github.com/filecoin-project/specs-actors/actors/runtime/exitcode"
 
 	. "github.com/filecoin-project/test-vectors/gen/builders"
+	. "github.com/filecoin-project/test-vectors/schema"
 )
 
 func main() {
