+++ gen/suites/vm_violations/main.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/filecoin-project/test-vectors/chaos"
 	. "github.com/filecoin-project/test-vectors/gen/builders"
+	. "github.com/filecoin-project/test-vectors/schema"
 )
 
 func main() {
