+++ schema/schema.go
@@ -1,4 +1,4 @@
-package builders
+package schema
 
 import (
 	"encoding/base64"
